refactor: 刪除 RecipeUploadStep1 元件以簡化結構，並更新引用路徑以改善檔案組織。
